moved router into its own folder. moved helpers into router folder. split out more helpers in router into helpers.
