+++ server/router/helpers/replacer.js
@@ -1,0 +1,12 @@
+const replacer = (key, value) => {
+  if (value && value.type && value.type === 'Buffer') {
+    return Buffer.from(value)
+      .toString()
+      .replace(/["]/g, '')
+      .trim();
+  }
+
+  return value;
+};
+
+module.exports = replacer;
